requirements.txt
django==5.0.2 
python-decouple==3.8 
psycopg2-binary==2.9.9 
djangorestframework==3.14.0 
django-cors-headers==4.3.1 
pydantic[email]==2.11.7
PyJWT==2.8.0
drf-spectacular==0.27.1
bcrypt==4.1.2
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
# Las librerías para email están incluidas en Python estándar (smtplib, email)    